Removed all references to serial ports
While I was unsure about what the purpose of serial ports was in the context of an IDE, I kept it around
in hopes of eventually getting it. It seems that it was just some code bloat with the sole purpose of
making it seem as if the IDE can work with embedded chips that can be connected to a PC's serial ports.
However, since the Run function is also a bloat that doesn't actually do anything, there is no longer
any need for any of them.

Signed-off-by: Robear Selwans <robear.selwans@outlook.com>